# deploy/helm/fuzekeys/files/openapi.yaml
...
openapi: 3.1.0
info:
  title: FuzeKeys API
  description: 'FuzeKeys — identity and credential vault. Managed digital identities, per-site accounts,
    encrypted vault assets, capability-token secret brokering, and automated signup workflows.


    DERIVED DOCUMENT: regenerate with `python scripts/export-openapi.py`. Do not hand-edit.'
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 2.0.0
paths:
...
  /api/v1/identities/:
    get:
      tags:
      - Identities
      summary: List Identities
      description: List all identities for the current user.
      operationId: get_identities
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/IdentityListResponse'
                type: array
                title: Response List Identities Api V1 Identities  Get
      security:
      - HTTPBearer: []
    post:
      tags:
      - Identities
      summary: Create Identity
      description: Create a new identity.
      operationId: post_identities
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdentityCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/identities/{identity_id}:
    get:
      tags:
      - Identities
      summary: Get Identity
      description: Get a specific identity.
      operationId: get_identities_by_identity_id
      security:
      - HTTPBearer: []
      parameters:
      - name: identity_id
        in: path
        required: true
        schema:
          type: integer
          title: Identity Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Identities
      summary: Update Identity
      description: Update an identity.
      operationId: put_identities_by_identity_id
      security:
      - HTTPBearer: []
      parameters:
      - name: identity_id
        in: path
        required: true
        schema:
          type: integer
          title: Identity Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdentityUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Identities
      summary: Delete Identity
      description: Delete an identity.
      operationId: delete_identities_by_identity_id
      security:
      - HTTPBearer: []
      parameters:
      - name: identity_id
        in: path
        required: true
        schema:
          type: integer
          title: Identity Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    IdentityCreate:
      properties:
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        first_name:
          type: string
          title: First Name
        last_name:
          type: string
          title: Last Name
        email:
          type: string
          format: email
          title: Email
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        date_of_birth:
          anyOf:
          - type: string
          - type: 'null'
          title: Date Of Birth
        address_line1:
          anyOf:
          - type: string
          - type: 'null'
          title: Address Line1
        address_line2:
          anyOf:
          - type: string
          - type: 'null'
          title: Address Line2
        city:
          anyOf:
          - type: string
          - type: 'null'
          title: City
        state:
          anyOf:
          - type: string
          - type: 'null'
          title: State
        zip_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Zip Code
        country:
          anyOf:
          - type: string
          - type: 'null'
          title: Country
        profession:
          anyOf:
          - type: string
          - type: 'null'
          title: Profession
        company:
          anyOf:
          - type: string
          - type: 'null'
          title: Company
        bio:
          anyOf:
          - type: string
          - type: 'null'
          title: Bio
        custom_fields:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Custom Fields
        preferred_username_pattern:
          anyOf:
          - type: string
          - type: 'null'
          title: Preferred Username Pattern
      additionalProperties: false
      type: object
      required:
      - name
      - first_name
      - last_name
      - email
      title: IdentityCreate
    IdentityListResponse:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - name
      - description
      - created_at
      title: IdentityListResponse
    IdentityResponse:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        first_name:
          type: string
          title: First Name
        last_name:
          type: string
          title: Last Name
        email:
          type: string
          title: Email
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        date_of_birth:
          anyOf:
          - type: string
          - type: 'null'
          title: Date Of Birth
        address_line1:
          anyOf:
          - type: string
          - type: 'null'
          title: Address Line1
        address_line2:
          anyOf:
          - type: string
          - type: 'null'
          title: Address Line2
        city:
          anyOf:
          - type: string
          - type: 'null'
          title: City
        state:
          anyOf:
          - type: string
          - type: 'null'
          title: State
        zip_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Zip Code
        country:
          anyOf:
          - type: string
          - type: 'null'
          title: Country
        profession:
          anyOf:
          - type: string
          - type: 'null'
          title: Profession
        company:
          anyOf:
          - type: string
          - type: 'null'
          title: Company
        bio:
          anyOf:
          - type: string
          - type: 'null'
          title: Bio
        custom_fields:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Custom Fields
        preferred_username_pattern:
          anyOf:
          - type: string
          - type: 'null'
          title: Preferred Username Pattern
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
      type: object
      required:
      - id
      - name
      - description
      - first_name
      - last_name
      - email
      - phone
      - date_of_birth
      - address_line1
      - address_line2
      - city
      - state
      - zip_code
      - country
      - profession
      - company
      - bio
      - custom_fields
      - preferred_username_pattern
      - created_at
      - updated_at
      title: IdentityResponse
    IdentityUpdate:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        first_name:
          anyOf:
          - type: string
          - type: 'null'
          title: First Name
        last_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Name
        email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Email
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        date_of_birth:
          anyOf:
          - type: string
          - type: 'null'
          title: Date Of Birth
        address_line1:
          anyOf:
          - type: string
          - type: 'null'
          title: Address Line1
        address_line2:
          anyOf:
          - type: string
          - type: 'null'
          title: Address Line2
        city:
          anyOf:
          - type: string
          - type: 'null'
          title: City
        state:
          anyOf:
          - type: string
          - type: 'null'
          title: State
        zip_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Zip Code
        country:
          anyOf:
          - type: string
          - type: 'null'
          title: Country
        profession:
          anyOf:
          - type: string
          - type: 'null'
          title: Profession
        company:
          anyOf:
          - type: string
          - type: 'null'
          title: Company
        bio:
          anyOf:
          - type: string
          - type: 'null'
          title: Bio
        custom_fields:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Custom Fields
        preferred_username_pattern:
          anyOf:
          - type: string
          - type: 'null'
          title: Preferred Username Pattern
      type: object
      title: IdentityUpdate
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
...